move servers out
